requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
gunicorn==23.0.0
jinja2==3.1.6
python-multipart==0.0.20
starlette==0.46.2
itsdangerous==2.2.0
pyotp==2.9.0
qrcode[pil]==8.2
pillow==11.2.1
python-dotenv==1.1.0
